Install docker without using Vagrant's provisioner
If the docker provisioner is used, Vagrant starts docker on up/reload
before it mounts synced folders. Docker ends up writing to the mount
point before the mount is ready, and then the mount complaints that the
mount point is not empty. Starting docker using vagrant-triggers allows
for the mount to be ready first.