Add loopback IP address for macOS devices
The idea was that this would allow for the docker containers to access
the VM's host machine (i.e. macOS) directly for databases and such that
are setup outside of the VM or a container on the VM.

I originally did this manually on another computer. After I got done
automating, I am realizing I think this can all be avoided by using
the host machine's IP address on the VM network: 10.0.2.2

I will revert this change immediately after committing it, but I want
the change around for the sake of history.